402 refactor: an app's chart lives with the app
apps/auth/deploy/pulumi and apps/files/deploy/pulumi, matching the shape
mcp-gateway and mariastew took on the way out. The chart pins the app's own
version and the two release together, so one directory holds both — and an app
leaving is then a subtree split of that directory rather than a hunt across two
trees.

serve-from-env had no chart to move and turned out to have one anyway: the
Secret, the Deployment, the routes-hash annotation that restarts it when the
table changes, and the hardened securityContext were all inline in
createProfileServer, mixed into building sing-box profiles. They are
serve-from-env's own shape and live with it now. What stays in @jaritanet/vpn
is what the table *means* — how a path is derived and what the body says —
which took profiles.ts from 171 lines to 118.

The line that came out of this: apps/ is source this repository builds, and
packages/ is deploy code for software built somewhere else, plus the two shared
primitives. Nothing in packages/ can be extracted, because there is nowhere to
extract it to.
