+++ apps/files/deploy/pulumi/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": ["../../../../tsconfig.json"],
+  "include": ["src/**/*.ts"]
+}
