+++ apps/serve-from-env/deploy/pulumi/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@jaritanet/serve-from-env",
+  "version": "0.0.0",
+  "private": true,
+  "description": "Serves a JSON table of path → body, and nothing else",
+  "license": "ISC",
+  "author": "James Cleveland <jc@blit.cc>",
+  "files": [
+    "src"
+  ],
+  "type": "module",
+  "main": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "dependencies": {
+    "@jaritanet/k8s": "workspace:*",
+    "@pulumi/kubernetes": "4.30.0",
+    "@pulumi/pulumi": "3.239.0"
+  }
+}
