+++ apps/serve-from-env/deploy/pulumi/src/index.ts
@@ -1,0 +1,2 @@
+export { createServeFromEnv } from "./serve-from-env.ts";
+export { VERSIONS as SERVE_FROM_ENV_VERSIONS } from "./versions.ts";
