+++ aube-lock.yaml
@@ -43,6 +43,33 @@ importers:
         specifier: 4.1.6
         version: 4.1.6(@opentelemetry/api@1.9.1)(@types/node@25.8.0)(@vitest/ui@4.1.6)(vite@8.1.5(@types/node@25.8.0))
 
+  apps/auth/deploy/pulumi:
+    dependencies:
+      '@jaritanet/k8s':
+        specifier: workspace:*
+        version: 0.0.0
+      '@pulumi/kubernetes':
+        specifier: 4.30.0
+        version: 4.30.0(typescript@6.0.3)
+      '@pulumi/pulumi':
+        specifier: 3.239.0
+        version: 3.239.0(typescript@6.0.3)
+      '@pulumi/random':
+        specifier: 4.20.0
+        version: 4.20.0(typescript@6.0.3)
+      zod:
+        specifier: 4.4.3
+        version: 4.4.3
+
+  apps/files/deploy/pulumi:
+    dependencies:
+      '@jaritanet/k8s':
+        specifier: workspace:*
+        version: 0.0.0
+      '@pulumi/kubernetes':
+        specifier: 4.30.0
+        version: 4.30.0(typescript@6.0.3)
+
   apps/mariastew:
     devDependencies:
       '@tailwindcss/cli':
@@ -64,7 +91,7 @@ importers:
         specifier: 6.0.3
         version: 6.0.3
 
-  packages/auth:
+  apps/serve-from-env/deploy/pulumi:
     dependencies:
       '@jaritanet/k8s':
         specifier: workspace:*
@@ -75,12 +102,6 @@ importers:
       '@pulumi/pulumi':
         specifier: 3.239.0
         version: 3.239.0(typescript@6.0.3)
-      '@pulumi/random':
-        specifier: 4.20.0
-        version: 4.20.0(typescript@6.0.3)
-      zod:
-        specifier: 4.4.3
-        version: 4.4.3
 
   packages/blit:
     dependencies:
@@ -105,15 +126,6 @@ importers:
       zod:
         specifier: 4.4.3
         version: 4.4.3
-
-  packages/files:
-    dependencies:
-      '@jaritanet/k8s':
-        specifier: workspace:*
-        version: 0.0.0
-      '@pulumi/kubernetes':
-        specifier: 4.30.0
-        version: 4.30.0(typescript@6.0.3)
 
   packages/hetzner:
     dependencies:
@@ -306,6 +318,9 @@ importers:
         specifier: workspace:*
         version: 0.0.0
       '@jaritanet/remote':
+        specifier: workspace:*
+        version: 0.0.0
+      '@jaritanet/serve-from-env':
         specifier: workspace:*
         version: 0.0.0
       '@pulumi/command':
