+++ packages/vpn/package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "@jaritanet/k8s": "workspace:*",
     "@jaritanet/remote": "workspace:*",
+    "@jaritanet/serve-from-env": "workspace:*",
     "@pulumi/command": "1.2.1",
     "@pulumi/kubernetes": "4.30.0",
     "@pulumi/pulumi": "3.239.0",
